+++ hi-ml-azure/src/health/azure/himl.py
@@ -68,7 +68,7 @@ class AzureRunInfo:
     run: Optional[Run]
     """An AzureML Run object if the present script is executing inside AzureML, or None if outside of AzureML.
     The Run object has methods to log metrics, upload files, etc."""
-    is_running_in_azure: bool
+    is_running_in_azure_ml: bool
     """If True, the present script is executing inside AzureML. If False, outside AzureML."""
     output_folder: Path
     """The output folder into which all script outputs should be written, if they should be later available in the
@@ -362,7 +362,7 @@ def submit_to_azure_if_needed(  # type: ignore
     # The present function will most likely be called from the script once it is running in AzureML.
     # The '--azureml' flag will not be present anymore, but we don't want to rely on that. From Run.get_context we
     # can infer if the present code is running in AzureML.
-    in_azure = is_running_in_azure_ml()
+    in_azure = is_running_in_azure_ml(RUN_CONTEXT)
     if in_azure:
         return _generate_azure_datasets(cleaned_input_datasets, cleaned_output_datasets)
 
@@ -386,7 +386,7 @@ def submit_to_azure_if_needed(  # type: ignore
             input_datasets=[d.local_folder for d in cleaned_input_datasets],
             output_datasets=[d.local_folder for d in cleaned_output_datasets],
             run=None,
-            is_running_in_azure=False,
+            is_running_in_azure_ml=False,
             output_folder=output_folder,
             logs_folder=logs_folder
         )
@@ -512,7 +512,7 @@ def _generate_azure_datasets(
         input_datasets=returned_input_datasets,  # type: ignore
         output_datasets=returned_output_datasets,  # type: ignore
         run=RUN_CONTEXT,
-        is_running_in_azure=True,
+        is_running_in_azure_ml=True,
         output_folder=Path.cwd() / OUTPUT_FOLDER,
         logs_folder=Path.cwd() / LOGS_FOLDER)